firstproject/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
requests==2.32.3 